+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+- Adding new externalRuntimeId, allows to pass external runtimeId to ChatSDK and keep in sync sessions for telemetry.
 - Add ability to use `CoreServices` orgUrl dynamically at runtime
 
 ## [1.8.0] - 2024-03-29